Avoid false-fails when the application under test exists quickly
If a program reads a sequence, that makes it terminate and `programexit'
is used, it might happen, that the application is already dead when the
unexpected_death() test is done. That condition is okay if the current
step in the test-script is `programexit', because that is exactly what
has happened.

The `programexit' handled will take care of checking the application's
exit-code.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>